// src/interfaces/Attributes.ts
export interface Attributes {
    hp: number;      // Punkty życia
    maxHp: number;
    mp: number;      // Punkty many
    maxMp: number;
    speed: number;   // Prędkość (inicjatywa w turze / punkty ruchu)
    strength: number; // Siła (obrażenia fizyczne)
    dexterity: number; // Zręczność (celność / obrażenia dystansowe)
    intelligence: number; // Inteligencja (moc zaklęć / mana)
}